rm resource-card, move resource-navigation and styled-resource-type to new files
