+++ ui/src/app/components/resources/resource-list.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { CSSProperties, useEffect, useState } from 'react';
-import { Link as RouterLink } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import * as fp from 'lodash/fp';
 import { Column } from 'primereact/column';
 import { DataTable } from 'primereact/datatable';
@@ -16,10 +16,6 @@ import {
 
 import { Clickable } from 'app/components/buttons';
 import { TooltipTrigger } from 'app/components/popups';
-import {
-  ResourceNavigation,
-  StyledResourceType,
-} from 'app/components/resource-card';
 import { renderResourceMenu } from 'app/components/resources/render-resource-menu';
 import { analysisTabPath, dataTabPath } from 'app/routing/utils';
 import colors from 'app/styles/colors';
@@ -34,6 +30,9 @@ import {
   isNotebook,
 } from 'app/utils/resources';
 import { WorkspaceData } from 'app/utils/workspace-data';
+
+import { ResourceNavigation } from './resource-navigation';
+import { StyledResourceType } from './styled-resource-type';
 
 const styles = reactStyles({
   column: {
@@ -77,9 +76,9 @@ const WorkspaceNavigation = (props: NavProps) => {
     : dataTabPath(namespace, id);
   return (
     <Clickable>
-      <RouterLink to={url} style={style} data-test-id='workspace-navigation'>
+      <Link to={url} style={style} data-test-id='workspace-navigation'>
         {name}
-      </RouterLink>
+      </Link>
     </Clickable>
   );
 };
